clearer deprecation boundary
